Module aws_sdk_opensearch::client::fluent_builders

Utilities to ergonomically construct a request to the service.

Fluent builders are created through the Client by calling one if its operation methods. After parameters are set using the builder methods, the send method can be called to initiate the request.

Structs

AcceptInboundConnection

Fluent builder constructing a request to AcceptInboundConnection.

AddTags

Fluent builder constructing a request to AddTags.

AssociatePackage

Fluent builder constructing a request to AssociatePackage.

CancelServiceSoftwareUpdate

Fluent builder constructing a request to CancelServiceSoftwareUpdate.

CreateDomain

Fluent builder constructing a request to CreateDomain.

CreateOutboundConnection

Fluent builder constructing a request to CreateOutboundConnection.

CreatePackage

Fluent builder constructing a request to CreatePackage.

DeleteDomain

Fluent builder constructing a request to DeleteDomain.

DeleteInboundConnection

Fluent builder constructing a request to DeleteInboundConnection.

DeleteOutboundConnection

Fluent builder constructing a request to DeleteOutboundConnection.

DeletePackage

Fluent builder constructing a request to DeletePackage.

DescribeDomain

Fluent builder constructing a request to DescribeDomain.

DescribeDomainAutoTunes

Fluent builder constructing a request to DescribeDomainAutoTunes.

DescribeDomainConfig

Fluent builder constructing a request to DescribeDomainConfig.

DescribeDomains

Fluent builder constructing a request to DescribeDomains.

DescribeInboundConnections

Fluent builder constructing a request to DescribeInboundConnections.

DescribeInstanceTypeLimits

Fluent builder constructing a request to DescribeInstanceTypeLimits.

DescribeOutboundConnections

Fluent builder constructing a request to DescribeOutboundConnections.

DescribePackages

Fluent builder constructing a request to DescribePackages.

DescribeReservedInstanceOfferings

Fluent builder constructing a request to DescribeReservedInstanceOfferings.

DescribeReservedInstances

Fluent builder constructing a request to DescribeReservedInstances.

DissociatePackage

Fluent builder constructing a request to DissociatePackage.

GetCompatibleVersions

Fluent builder constructing a request to GetCompatibleVersions.

GetPackageVersionHistory

Fluent builder constructing a request to GetPackageVersionHistory.

GetUpgradeHistory

Fluent builder constructing a request to GetUpgradeHistory.

GetUpgradeStatus

Fluent builder constructing a request to GetUpgradeStatus.

ListDomainNames

Fluent builder constructing a request to ListDomainNames.

ListDomainsForPackage

Fluent builder constructing a request to ListDomainsForPackage.

ListInstanceTypeDetails

Fluent builder constructing a request to ListInstanceTypeDetails.

ListPackagesForDomain

Fluent builder constructing a request to ListPackagesForDomain.

ListTags

Fluent builder constructing a request to ListTags.

ListVersions

Fluent builder constructing a request to ListVersions.

PurchaseReservedInstanceOffering

Fluent builder constructing a request to PurchaseReservedInstanceOffering.

RejectInboundConnection

Fluent builder constructing a request to RejectInboundConnection.

RemoveTags

Fluent builder constructing a request to RemoveTags.

StartServiceSoftwareUpdate

Fluent builder constructing a request to StartServiceSoftwareUpdate.

UpdateDomainConfig

Fluent builder constructing a request to UpdateDomainConfig.

UpdatePackage

Fluent builder constructing a request to UpdatePackage.

UpgradeDomain

Fluent builder constructing a request to UpgradeDomain.