Struct aws_sdk_opensearch::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for AmazonOpenSearchService.

This client allows ergonomic access to a AmazonOpenSearchService-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

Creates a client with the given service configuration.

pub fn conf(&self) -> &Config

Returns the client’s configuration.

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn accept_inbound_connection(&self) -> AcceptInboundConnection<C, M, R>

Constructs a fluent builder for the AcceptInboundConnection operation.

See AcceptInboundConnection for more information about the operation and its arguments.

pub fn add_tags(&self) -> AddTags<C, M, R>

Constructs a fluent builder for the AddTags operation.

See AddTags for more information about the operation and its arguments.

pub fn associate_package(&self) -> AssociatePackage<C, M, R>

Constructs a fluent builder for the AssociatePackage operation.

See AssociatePackage for more information about the operation and its arguments.

pub fn cancel_service_software_update(
&self
) -> CancelServiceSoftwareUpdate<C, M, R>

Constructs a fluent builder for the CancelServiceSoftwareUpdate operation.

See CancelServiceSoftwareUpdate for more information about the operation and its arguments.

pub fn create_domain(&self) -> CreateDomain<C, M, R>

Constructs a fluent builder for the CreateDomain operation.

See CreateDomain for more information about the operation and its arguments.

pub fn create_outbound_connection(&self) -> CreateOutboundConnection<C, M, R>

Constructs a fluent builder for the CreateOutboundConnection operation.

See CreateOutboundConnection for more information about the operation and its arguments.

pub fn create_package(&self) -> CreatePackage<C, M, R>

Constructs a fluent builder for the CreatePackage operation.

See CreatePackage for more information about the operation and its arguments.

pub fn delete_domain(&self) -> DeleteDomain<C, M, R>

Constructs a fluent builder for the DeleteDomain operation.

See DeleteDomain for more information about the operation and its arguments.

pub fn delete_inbound_connection(&self) -> DeleteInboundConnection<C, M, R>

Constructs a fluent builder for the DeleteInboundConnection operation.

See DeleteInboundConnection for more information about the operation and its arguments.

pub fn delete_outbound_connection(&self) -> DeleteOutboundConnection<C, M, R>

Constructs a fluent builder for the DeleteOutboundConnection operation.

See DeleteOutboundConnection for more information about the operation and its arguments.

pub fn delete_package(&self) -> DeletePackage<C, M, R>

Constructs a fluent builder for the DeletePackage operation.

See DeletePackage for more information about the operation and its arguments.

pub fn describe_domain(&self) -> DescribeDomain<C, M, R>

Constructs a fluent builder for the DescribeDomain operation.

See DescribeDomain for more information about the operation and its arguments.

pub fn describe_domain_auto_tunes(&self) -> DescribeDomainAutoTunes<C, M, R>

Constructs a fluent builder for the DescribeDomainAutoTunes operation.

See DescribeDomainAutoTunes for more information about the operation and its arguments.

pub fn describe_domain_config(&self) -> DescribeDomainConfig<C, M, R>

Constructs a fluent builder for the DescribeDomainConfig operation.

See DescribeDomainConfig for more information about the operation and its arguments.

pub fn describe_domains(&self) -> DescribeDomains<C, M, R>

Constructs a fluent builder for the DescribeDomains operation.

See DescribeDomains for more information about the operation and its arguments.

pub fn describe_inbound_connections(
&self
) -> DescribeInboundConnections<C, M, R>

Constructs a fluent builder for the DescribeInboundConnections operation.

See DescribeInboundConnections for more information about the operation and its arguments.

pub fn describe_instance_type_limits(
&self
) -> DescribeInstanceTypeLimits<C, M, R>

Constructs a fluent builder for the DescribeInstanceTypeLimits operation.

See DescribeInstanceTypeLimits for more information about the operation and its arguments.

pub fn describe_outbound_connections(
&self
) -> DescribeOutboundConnections<C, M, R>

Constructs a fluent builder for the DescribeOutboundConnections operation.

See DescribeOutboundConnections for more information about the operation and its arguments.

pub fn describe_packages(&self) -> DescribePackages<C, M, R>

Constructs a fluent builder for the DescribePackages operation.

See DescribePackages for more information about the operation and its arguments.

pub fn describe_reserved_instance_offerings(
&self
) -> DescribeReservedInstanceOfferings<C, M, R>

Constructs a fluent builder for the DescribeReservedInstanceOfferings operation.

See DescribeReservedInstanceOfferings for more information about the operation and its arguments.

pub fn describe_reserved_instances(&self) -> DescribeReservedInstances<C, M, R>

Constructs a fluent builder for the DescribeReservedInstances operation.

See DescribeReservedInstances for more information about the operation and its arguments.

pub fn dissociate_package(&self) -> DissociatePackage<C, M, R>

Constructs a fluent builder for the DissociatePackage operation.

See DissociatePackage for more information about the operation and its arguments.

pub fn get_compatible_versions(&self) -> GetCompatibleVersions<C, M, R>

Constructs a fluent builder for the GetCompatibleVersions operation.

See GetCompatibleVersions for more information about the operation and its arguments.

pub fn get_package_version_history(&self) -> GetPackageVersionHistory<C, M, R>

Constructs a fluent builder for the GetPackageVersionHistory operation.

See GetPackageVersionHistory for more information about the operation and its arguments.

pub fn get_upgrade_history(&self) -> GetUpgradeHistory<C, M, R>

Constructs a fluent builder for the GetUpgradeHistory operation.

See GetUpgradeHistory for more information about the operation and its arguments.

pub fn get_upgrade_status(&self) -> GetUpgradeStatus<C, M, R>

Constructs a fluent builder for the GetUpgradeStatus operation.

See GetUpgradeStatus for more information about the operation and its arguments.

pub fn list_domain_names(&self) -> ListDomainNames<C, M, R>

Constructs a fluent builder for the ListDomainNames operation.

See ListDomainNames for more information about the operation and its arguments.

pub fn list_domains_for_package(&self) -> ListDomainsForPackage<C, M, R>

Constructs a fluent builder for the ListDomainsForPackage operation.

See ListDomainsForPackage for more information about the operation and its arguments.

pub fn list_instance_type_details(&self) -> ListInstanceTypeDetails<C, M, R>

Constructs a fluent builder for the ListInstanceTypeDetails operation.

See ListInstanceTypeDetails for more information about the operation and its arguments.

pub fn list_packages_for_domain(&self) -> ListPackagesForDomain<C, M, R>

Constructs a fluent builder for the ListPackagesForDomain operation.

See ListPackagesForDomain for more information about the operation and its arguments.

pub fn list_tags(&self) -> ListTags<C, M, R>

Constructs a fluent builder for the ListTags operation.

See ListTags for more information about the operation and its arguments.

pub fn list_versions(&self) -> ListVersions<C, M, R>

Constructs a fluent builder for the ListVersions operation.

See ListVersions for more information about the operation and its arguments.

pub fn purchase_reserved_instance_offering(
&self
) -> PurchaseReservedInstanceOffering<C, M, R>

Constructs a fluent builder for the PurchaseReservedInstanceOffering operation.

See PurchaseReservedInstanceOffering for more information about the operation and its arguments.

pub fn reject_inbound_connection(&self) -> RejectInboundConnection<C, M, R>

Constructs a fluent builder for the RejectInboundConnection operation.

See RejectInboundConnection for more information about the operation and its arguments.

pub fn remove_tags(&self) -> RemoveTags<C, M, R>

Constructs a fluent builder for the RemoveTags operation.

See RemoveTags for more information about the operation and its arguments.

pub fn start_service_software_update(
&self
) -> StartServiceSoftwareUpdate<C, M, R>

Constructs a fluent builder for the StartServiceSoftwareUpdate operation.

See StartServiceSoftwareUpdate for more information about the operation and its arguments.

pub fn update_domain_config(&self) -> UpdateDomainConfig<C, M, R>

Constructs a fluent builder for the UpdateDomainConfig operation.

See UpdateDomainConfig for more information about the operation and its arguments.

pub fn update_package(&self) -> UpdatePackage<C, M, R>

Constructs a fluent builder for the UpdatePackage operation.

See UpdatePackage for more information about the operation and its arguments.

pub fn upgrade_domain(&self) -> UpgradeDomain<C, M, R>

Constructs a fluent builder for the UpgradeDomain operation.

See UpgradeDomain for more information about the operation and its arguments.

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

Creates a client with the given service config and connector override.

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

Creates a new client from a shared config.

pub fn from_conf(conf: Config) -> Self

Creates a new client from the service Config.