Enum aws_sdk_opensearch::Error

#[non_exhaustive]
pub enum Error {
    AccessDeniedException(AccessDeniedException),
    BaseException(BaseException),
    ConflictException(ConflictException),
    DisabledOperationException(DisabledOperationException),
    InternalException(InternalException),
    InvalidPaginationTokenException(InvalidPaginationTokenException),
    InvalidTypeException(InvalidTypeException),
    LimitExceededException(LimitExceededException),
    ResourceAlreadyExistsException(ResourceAlreadyExistsException),
    ResourceNotFoundException(ResourceNotFoundException),
    ValidationException(ValidationException),
    Unhandled(Box<dyn Error + Send + Sync + 'static>),
}

All possible error types for this service.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

AccessDeniedException(AccessDeniedException)

An error occurred because user does not have permissions to access the resource. Returns HTTP status code 403.

Tuple Fields of AccessDeniedException

0: AccessDeniedException

BaseException(BaseException)

An error occurred while processing the request.

Tuple Fields of BaseException

0: BaseException

ConflictException(ConflictException)

An error occurred because the client attempts to remove a resource that is currently in use. Returns HTTP status code 409.

Tuple Fields of ConflictException

0: ConflictException

DisabledOperationException(DisabledOperationException)

An error occured because the client wanted to access a not supported operation. Gives http status code of 409.

Tuple Fields of DisabledOperationException

0: DisabledOperationException

InternalException(InternalException)

The request processing has failed because of an unknown error, exception or failure (the failure is internal to the service) . Gives http status code of 500.

Tuple Fields of InternalException

0: InternalException

InvalidPaginationTokenException(InvalidPaginationTokenException)

The request processing has failed because of invalid pagination token provided by customer. Returns an HTTP status code of 400.

Tuple Fields of InvalidPaginationTokenException

0: InvalidPaginationTokenException

InvalidTypeException(InvalidTypeException)

An exception for trying to create or access sub-resource that is either invalid or not supported. Gives http status code of 409.

Tuple Fields of InvalidTypeException

0: InvalidTypeException

LimitExceededException(LimitExceededException)

An exception for trying to create more than allowed resources or sub-resources. Gives http status code of 409.

Tuple Fields of LimitExceededException

0: LimitExceededException

ResourceAlreadyExistsException(ResourceAlreadyExistsException)

An exception for creating a resource that already exists. Gives http status code of 400.

Tuple Fields of ResourceAlreadyExistsException

0: ResourceAlreadyExistsException

ResourceNotFoundException(ResourceNotFoundException)

An exception for accessing or deleting a resource that does not exist. Gives http status code of 400.

Tuple Fields of ResourceNotFoundException

0: ResourceNotFoundException

ValidationException(ValidationException)

An exception for missing / invalid input fields. Gives http status code of 400.

Tuple Fields of ValidationException

0: ValidationException

Unhandled(Box<dyn Error + Send + Sync + 'static>)

An unhandled error occurred.

Tuple Fields of Unhandled

0: Box<dyn Error + Send + Sync + 'static>