Module aws_sdk_opensearch::error

Errors that can occur when calling the service.

Modules

access_denied_exception

See AccessDeniedException

base_exception

See BaseException

conflict_exception

See ConflictException

disabled_operation_exception

See DisabledOperationException

internal_exception

See InternalException

invalid_pagination_token_exception

See InvalidPaginationTokenException

invalid_type_exception

See InvalidTypeException

limit_exceeded_exception

See LimitExceededException

resource_already_exists_exception

See ResourceAlreadyExistsException

resource_not_found_exception

See ResourceNotFoundException

validation_exception

See ValidationException

Structs

AcceptInboundConnectionError

Error type for the AcceptInboundConnection operation.

AccessDeniedException

An error occurred because user does not have permissions to access the resource. Returns HTTP status code 403.

AddTagsError

Error type for the AddTags operation.

AssociatePackageError

Error type for the AssociatePackage operation.

BaseException

An error occurred while processing the request.

CancelServiceSoftwareUpdateError

Error type for the CancelServiceSoftwareUpdate operation.

ConflictException

An error occurred because the client attempts to remove a resource that is currently in use. Returns HTTP status code 409.

CreateDomainError

Error type for the CreateDomain operation.

CreateOutboundConnectionError

Error type for the CreateOutboundConnection operation.

CreatePackageError

Error type for the CreatePackage operation.

DeleteDomainError

Error type for the DeleteDomain operation.

DeleteInboundConnectionError

Error type for the DeleteInboundConnection operation.

DeleteOutboundConnectionError

Error type for the DeleteOutboundConnection operation.

DeletePackageError

Error type for the DeletePackage operation.

DescribeDomainAutoTunesError

Error type for the DescribeDomainAutoTunes operation.

DescribeDomainConfigError

Error type for the DescribeDomainConfig operation.

DescribeDomainError

Error type for the DescribeDomain operation.

DescribeDomainsError

Error type for the DescribeDomains operation.

DescribeInboundConnectionsError

Error type for the DescribeInboundConnections operation.

DescribeInstanceTypeLimitsError

Error type for the DescribeInstanceTypeLimits operation.

DescribeOutboundConnectionsError

Error type for the DescribeOutboundConnections operation.

DescribePackagesError

Error type for the DescribePackages operation.

DescribeReservedInstanceOfferingsError

Error type for the DescribeReservedInstanceOfferings operation.

DescribeReservedInstancesError

Error type for the DescribeReservedInstances operation.

DisabledOperationException

An error occured because the client wanted to access a not supported operation. Gives http status code of 409.

DissociatePackageError

Error type for the DissociatePackage operation.

GetCompatibleVersionsError

Error type for the GetCompatibleVersions operation.

GetPackageVersionHistoryError

Error type for the GetPackageVersionHistory operation.

GetUpgradeHistoryError

Error type for the GetUpgradeHistory operation.

GetUpgradeStatusError

Error type for the GetUpgradeStatus operation.

InternalException

The request processing has failed because of an unknown error, exception or failure (the failure is internal to the service) . Gives http status code of 500.

InvalidPaginationTokenException

The request processing has failed because of invalid pagination token provided by customer. Returns an HTTP status code of 400.

InvalidTypeException

An exception for trying to create or access sub-resource that is either invalid or not supported. Gives http status code of 409.

LimitExceededException

An exception for trying to create more than allowed resources or sub-resources. Gives http status code of 409.

ListDomainNamesError

Error type for the ListDomainNames operation.

ListDomainsForPackageError

Error type for the ListDomainsForPackage operation.

ListInstanceTypeDetailsError

Error type for the ListInstanceTypeDetails operation.

ListPackagesForDomainError

Error type for the ListPackagesForDomain operation.

ListTagsError

Error type for the ListTags operation.

ListVersionsError

Error type for the ListVersions operation.

PurchaseReservedInstanceOfferingError

Error type for the PurchaseReservedInstanceOffering operation.

RejectInboundConnectionError

Error type for the RejectInboundConnection operation.

RemoveTagsError

Error type for the RemoveTags operation.

ResourceAlreadyExistsException

An exception for creating a resource that already exists. Gives http status code of 400.

ResourceNotFoundException

An exception for accessing or deleting a resource that does not exist. Gives http status code of 400.

StartServiceSoftwareUpdateError

Error type for the StartServiceSoftwareUpdate operation.

UpdateDomainConfigError

Error type for the UpdateDomainConfig operation.

UpdatePackageError

Error type for the UpdatePackage operation.

UpgradeDomainError

Error type for the UpgradeDomain operation.

ValidationException

An exception for missing / invalid input fields. Gives http status code of 400.

Enums

AcceptInboundConnectionErrorKind

Types of errors that can occur for the AcceptInboundConnection operation.

AddTagsErrorKind

Types of errors that can occur for the AddTags operation.

AssociatePackageErrorKind

Types of errors that can occur for the AssociatePackage operation.

CancelServiceSoftwareUpdateErrorKind

Types of errors that can occur for the CancelServiceSoftwareUpdate operation.

CreateDomainErrorKind

Types of errors that can occur for the CreateDomain operation.

CreateOutboundConnectionErrorKind

Types of errors that can occur for the CreateOutboundConnection operation.

CreatePackageErrorKind

Types of errors that can occur for the CreatePackage operation.

DeleteDomainErrorKind

Types of errors that can occur for the DeleteDomain operation.

DeleteInboundConnectionErrorKind

Types of errors that can occur for the DeleteInboundConnection operation.

DeleteOutboundConnectionErrorKind

Types of errors that can occur for the DeleteOutboundConnection operation.

DeletePackageErrorKind

Types of errors that can occur for the DeletePackage operation.

DescribeDomainAutoTunesErrorKind

Types of errors that can occur for the DescribeDomainAutoTunes operation.

DescribeDomainConfigErrorKind

Types of errors that can occur for the DescribeDomainConfig operation.

DescribeDomainErrorKind

Types of errors that can occur for the DescribeDomain operation.

DescribeDomainsErrorKind

Types of errors that can occur for the DescribeDomains operation.

DescribeInboundConnectionsErrorKind

Types of errors that can occur for the DescribeInboundConnections operation.

DescribeInstanceTypeLimitsErrorKind

Types of errors that can occur for the DescribeInstanceTypeLimits operation.

DescribeOutboundConnectionsErrorKind

Types of errors that can occur for the DescribeOutboundConnections operation.

DescribePackagesErrorKind

Types of errors that can occur for the DescribePackages operation.

DescribeReservedInstanceOfferingsErrorKind

Types of errors that can occur for the DescribeReservedInstanceOfferings operation.

DescribeReservedInstancesErrorKind

Types of errors that can occur for the DescribeReservedInstances operation.

DissociatePackageErrorKind

Types of errors that can occur for the DissociatePackage operation.

GetCompatibleVersionsErrorKind

Types of errors that can occur for the GetCompatibleVersions operation.

GetPackageVersionHistoryErrorKind

Types of errors that can occur for the GetPackageVersionHistory operation.

GetUpgradeHistoryErrorKind

Types of errors that can occur for the GetUpgradeHistory operation.

GetUpgradeStatusErrorKind

Types of errors that can occur for the GetUpgradeStatus operation.

ListDomainNamesErrorKind

Types of errors that can occur for the ListDomainNames operation.

ListDomainsForPackageErrorKind

Types of errors that can occur for the ListDomainsForPackage operation.

ListInstanceTypeDetailsErrorKind

Types of errors that can occur for the ListInstanceTypeDetails operation.

ListPackagesForDomainErrorKind

Types of errors that can occur for the ListPackagesForDomain operation.

ListTagsErrorKind

Types of errors that can occur for the ListTags operation.

ListVersionsErrorKind

Types of errors that can occur for the ListVersions operation.

PurchaseReservedInstanceOfferingErrorKind

Types of errors that can occur for the PurchaseReservedInstanceOffering operation.

RejectInboundConnectionErrorKind

Types of errors that can occur for the RejectInboundConnection operation.

RemoveTagsErrorKind

Types of errors that can occur for the RemoveTags operation.

StartServiceSoftwareUpdateErrorKind

Types of errors that can occur for the StartServiceSoftwareUpdate operation.

UpdateDomainConfigErrorKind

Types of errors that can occur for the UpdateDomainConfig operation.

UpdatePackageErrorKind

Types of errors that can occur for the UpdatePackage operation.

UpgradeDomainErrorKind

Types of errors that can occur for the UpgradeDomain operation.