Module aws_sdk_opensearch::model

Data structures used by operation inputs/outputs.

Modules

access_policies_status

See AccessPoliciesStatus

additional_limit

See AdditionalLimit

advanced_options_status

See AdvancedOptionsStatus

advanced_security_options

See AdvancedSecurityOptions

advanced_security_options_input

See AdvancedSecurityOptionsInput

advanced_security_options_status

See AdvancedSecurityOptionsStatus

auto_tune

See AutoTune

auto_tune_details

See AutoTuneDetails

auto_tune_maintenance_schedule

See AutoTuneMaintenanceSchedule

auto_tune_options

See AutoTuneOptions

auto_tune_options_input

See AutoTuneOptionsInput

auto_tune_options_output

See AutoTuneOptionsOutput

auto_tune_options_status

See AutoTuneOptionsStatus

auto_tune_status

See AutoTuneStatus

aws_domain_information

See AwsDomainInformation

cluster_config

See ClusterConfig

cluster_config_status

See ClusterConfigStatus

cognito_options

See CognitoOptions

cognito_options_status

See CognitoOptionsStatus

cold_storage_options

See ColdStorageOptions

compatible_versions_map

See CompatibleVersionsMap

describe_packages_filter

See DescribePackagesFilter

domain_config

See DomainConfig

domain_endpoint_options

See DomainEndpointOptions

domain_endpoint_options_status

See DomainEndpointOptionsStatus

domain_info

See DomainInfo

domain_information_container

See DomainInformationContainer

domain_package_details

See DomainPackageDetails

domain_status

See DomainStatus

duration

See Duration

ebs_options

See EbsOptions

ebs_options_status

See EbsOptionsStatus

encryption_at_rest_options

See EncryptionAtRestOptions

encryption_at_rest_options_status

See EncryptionAtRestOptionsStatus

error_details

See ErrorDetails

filter

See Filter

inbound_connection

See InboundConnection

inbound_connection_status

See InboundConnectionStatus

instance_count_limits

See InstanceCountLimits

instance_limits

See InstanceLimits

instance_type_details

See InstanceTypeDetails

limits

See Limits

log_publishing_option

See LogPublishingOption

log_publishing_options_status

See LogPublishingOptionsStatus

master_user_options

See MasterUserOptions

node_to_node_encryption_options

See NodeToNodeEncryptionOptions

node_to_node_encryption_options_status

See NodeToNodeEncryptionOptionsStatus

option_status

See OptionStatus

outbound_connection

See OutboundConnection

outbound_connection_status

See OutboundConnectionStatus

package_details

See PackageDetails

package_source

See PackageSource

package_version_history

See PackageVersionHistory

recurring_charge

See RecurringCharge

reserved_instance

See ReservedInstance

reserved_instance_offering

See ReservedInstanceOffering

saml_idp

See SamlIdp

saml_options_input

See SamlOptionsInput

saml_options_output

See SamlOptionsOutput

scheduled_auto_tune_details

See ScheduledAutoTuneDetails

service_software_options

See ServiceSoftwareOptions

snapshot_options

See SnapshotOptions

snapshot_options_status

See SnapshotOptionsStatus

storage_type

See StorageType

storage_type_limit

See StorageTypeLimit

tag

See Tag

upgrade_history

See UpgradeHistory

upgrade_step_item

See UpgradeStepItem

version_status

See VersionStatus

vpc_derived_info

See VpcDerivedInfo

vpc_derived_info_status

See VpcDerivedInfoStatus

vpc_options

See VpcOptions

zone_awareness_config

See ZoneAwarenessConfig

Structs

AccessPoliciesStatus

The configured access rules for the domain's document and search endpoints, and the current status of those

AdditionalLimit

AdvancedOptionsStatus

Status of the advanced options for the specified domain. Currently, the following advanced options

AdvancedSecurityOptions

The advanced security configuration: whether advanced security is enabled, whether the internal

AdvancedSecurityOptionsInput

The advanced security configuration: whether advanced security is enabled, whether the internal

AdvancedSecurityOptionsStatus

The status of advanced security options for the specified domain.

AutoTune

Specifies the Auto-Tune type and Auto-Tune action details.

AutoTuneDetails

Specifies details about the Auto-Tune action. See

AutoTuneMaintenanceSchedule

Specifies the Auto-Tune maintenance schedule. See

AutoTuneOptions

The Auto-Tune options: the Auto-Tune desired state for the domain, rollback state when disabling

AutoTuneOptionsInput

The Auto-Tune options: the Auto-Tune desired state for the domain and list of maintenance schedules.

AutoTuneOptionsOutput

The Auto-Tune options: the Auto-Tune desired state for the domain and list of maintenance schedules.

AutoTuneOptionsStatus

The Auto-Tune status for the domain.

AutoTuneStatus

Provides the current Auto-Tune status for the domain.

AwsDomainInformation

ClusterConfig

The configuration for the domain cluster, such as the type and number of instances.

ClusterConfigStatus

The configuration status for the specified domain.

CognitoOptions

Options to specify the Cognito user and identity pools for OpenSearch Dashboards authentication. For more information, see Configuring Amazon Cognito authentication for OpenSearch Dashboards.

CognitoOptionsStatus

The status of the Cognito options for the specified domain.

ColdStorageOptions

Specifies the configuration for cold storage options such as enabled

CompatibleVersionsMap

DescribePackagesFilter

A filter to apply to the DescribePackage response.

DomainConfig

The configuration of a domain.

DomainEndpointOptions

Options to configure the endpoint for the domain.

DomainEndpointOptionsStatus

The configured endpoint options for the domain and their current status.

DomainInfo

DomainInformationContainer

DomainPackageDetails

Information on a package associated with a domain.

DomainStatus

The current status of a domain.

Duration

The maintenance schedule duration: duration value and duration unit. See

EbsOptions

Options to enable, disable, and specify the properties of EBS storage volumes.

EbsOptionsStatus

Status of the EBS options for the specified domain.

EncryptionAtRestOptions

Specifies encryption at rest options.

EncryptionAtRestOptionsStatus

Status of the encryption At Rest options for the specified domain.

ErrorDetails

Filter

InboundConnection

Details of an inbound connection.

InboundConnectionStatus

The connection status of an inbound cross-cluster connection.

InstanceCountLimits

InstanceLimits

InstanceLimits represents the list of instance-related attributes that are available for a given InstanceType.

InstanceTypeDetails

Limits

LogPublishingOption

Log Publishing option that is set for a given domain.

LogPublishingOptionsStatus

The configured log publishing options for the domain and their current status.

MasterUserOptions

Credentials for the master user: username and password, ARN, or both.

NodeToNodeEncryptionOptions

The node-to-node encryption options.

NodeToNodeEncryptionOptionsStatus

Status of the node-to-node encryption options for the specified domain.

OptionStatus

Provides the current status of the entity.

OutboundConnection

Specifies details about an outbound connection.

OutboundConnectionStatus

The connection status of an outbound cross-cluster connection.

PackageDetails

Basic information about a package.

PackageSource

The Amazon S3 location for importing the package specified as S3BucketName and

PackageVersionHistory

Details of a package version.

RecurringCharge

Contains the specific price and frequency of a recurring charges for a reserved OpenSearch instance, or for a

ReservedInstance

Details of a reserved OpenSearch instance.

ReservedInstanceOffering

Details of a reserved OpenSearch instance offering.

SamlIdp

The SAML identity povider's information.

SamlOptionsInput

The SAML application configuration for the domain.

SamlOptionsOutput

Describes the SAML application configured for the domain.

ScheduledAutoTuneDetails

Specifies details about the scheduled Auto-Tune action. See

ServiceSoftwareOptions

The current options of an domain service software options.

SnapshotOptions

The time, in UTC format, when the service takes a daily automated snapshot of the specified

SnapshotOptionsStatus

Status of a daily automated snapshot.

StorageType

StorageTypes represents the list of storage-related types and their attributes

StorageTypeLimit

Limits that are applicable for the given storage type.

Tag

A key value pair for a resource tag.

UpgradeHistory

History of the last 10 upgrades and upgrade eligibility checks.

UpgradeStepItem

Represents a single step of the upgrade or upgrade eligibility check workflow.

VersionStatus

The status of the OpenSearch version options for the specified OpenSearch domain.

VpcDerivedInfo

Options to specify the subnets and security groups for the VPC endpoint. For more information, see

VpcDerivedInfoStatus

Status of the VPC options for the specified domain.

VpcOptions

Options to specify the subnets and security groups for the VPC endpoint. For more information, see

ZoneAwarenessConfig

The zone awareness configuration for the domain cluster, such as the number of availability zones.

Enums

AutoTuneDesiredState

The Auto-Tune desired state. Valid values are ENABLED and DISABLED.

AutoTuneState

The Auto-Tune state for the domain. For valid states see

AutoTuneType

Specifies the Auto-Tune type. Valid value is SCHEDULED_ACTION.

DeploymentStatus

DescribePackagesFilterName

DomainPackageStatus

EngineType

InboundConnectionStatusCode

LogType

Type of log file. Can be one of the following:

OpenSearchPartitionInstanceType

OpenSearchWarmPartitionInstanceType

OptionState

The state of a requested change. One of the following:

OutboundConnectionStatusCode

PackageStatus

PackageType

ReservedInstancePaymentOption

RollbackOnDisable

The rollback state while disabling Auto-Tune for the domain. Valid values are NO_ROLLBACK and DEFAULT_ROLLBACK.

ScheduledAutoTuneActionType

The Auto-Tune action type. Valid values are JVM_HEAP_SIZE_TUNING, and JVM_YOUNG_GEN_TUNING.

ScheduledAutoTuneSeverityType

The Auto-Tune action severity. Valid values are LOW, MEDIUM, and HIGH.

TimeUnit

The unit of a maintenance schedule duration. Valid value is HOUR. See

TlsSecurityPolicy

UpgradeStatus

UpgradeStep

VolumeType

The type of EBS volume, standard, gp2, or io1. See Configuring EBS-based Storage for more information.